fix: thruthy renamed to  notnot
