+++ src/operations/logic/index.js
@@ -1,6 +1,6 @@
 export { default as equal } from './equal';
 export { default as not } from './not';
 export { default as notEqual } from './notEqual';
+export { default as notnot } from './notnot';
 export { default as strictEqual } from './strictEqual';
 export { default as strictNotEqual } from './strictNotEqual';
-export { default as truthy } from './truthy';
